refactor(identity): lighter account-syncing e2e tests
