+++ test/e2e/tests/identity/account-syncing/adding-and-renaming-accounts.spec.ts
@@ -1,0 +1,217 @@
+import { strict as assert } from 'assert';
+import { Mockttp } from 'mockttp';
+import { USER_STORAGE_FEATURE_NAMES } from '@metamask/profile-sync-controller/sdk';
+import { withFixtures, unlockWallet } from '../../../helpers';
+import { completeImportSRPOnboardingFlow } from '../../../page-objects/flows/onboarding.flow';
+import FixtureBuilder from '../../../fixture-builder';
+import { mockIdentityServices } from '../mocks';
+import { ACCOUNT_TYPE } from '../../../constants';
+import {
+  UserStorageMockttpController,
+  UserStorageMockttpControllerEvents,
+} from '../../../helpers/identity/user-storage/userStorageMockttpController';
+import HeaderNavbar from '../../../page-objects/pages/header-navbar';
+import AccountListPage from '../../../page-objects/pages/account-list-page';
+import AccountDetailsModal from '../../../page-objects/pages/dialog/account-details-modal';
+import { arrangeTestUtils } from './helpers';
+import { WALLET_PASSWORD } from '../../../helpers';
+import { E2E_SRP } from '../../../default-fixture';
+
+describe('Account syncing - Adding and Renaming Accounts', function () {
+  this.timeout(160000); // This test is very long, so we need an unusually high timeout
+
+  const DEFAULT_ACCOUNT_NAME = 'Account 1';
+  const ADDED_ACCOUNT_NAME = 'Account 2';
+  const NEW_ACCOUNT_NAME = 'RENAMED ACCOUNT';
+  const LAST_ACCOUNT_NAME = 'Account 3';
+
+  /**
+   * This test verifies the complete account syncing flow for adding and renaming accounts across three phases:
+   * Phase 1: From a loaded state with account syncing enabled, add a new account to the wallet and verify it syncs to user storage, checking that both the default account and newly added account are visible.
+   * Phase 2: Login to a fresh app instance, verify the previously added account persists, rename the second account, and add a third account to test multi-operation syncing.
+   * Phase 3: Complete onboarding flow from scratch to verify all account changes (additions and renames) are properly synced and persisted across app reinstallation.
+   */
+  it('should add a new account and sync it across multiple phases', async function () {
+    const userStorageMockttpController = new UserStorageMockttpController();
+
+    const sharedMockSetup = (server: Mockttp) => {
+      userStorageMockttpController.setupPath(
+        USER_STORAGE_FEATURE_NAMES.accounts,
+        server,
+      );
+      return mockIdentityServices(server, userStorageMockttpController);
+    };
+
+    // Phase 1: Add a new account and verify it syncs
+    await withFixtures(
+      {
+        fixtures: new FixtureBuilder()
+          .withBackupAndSyncSettings()
+          .build(),
+        title: this.test?.fullTitle(),
+        testSpecificMock: sharedMockSetup,
+      },
+      async ({ driver }) => {
+        await unlockWallet(driver);
+
+        const header = new HeaderNavbar(driver);
+        await header.check_pageIsLoaded();
+        await header.openAccountMenu();
+
+        const accountListPage = new AccountListPage(driver);
+        await accountListPage.check_pageIsLoaded();
+
+        // Verify default account is visible
+        await accountListPage.check_accountDisplayedInAccountList(
+          DEFAULT_ACCOUNT_NAME,
+        );
+
+        // Set up event counter to track sync operations
+        const { prepareEventsEmittedCounter, waitUntilSyncedAccountsNumberEquals } = arrangeTestUtils(
+          driver,
+          userStorageMockttpController,
+        );
+        const { waitUntilEventsEmittedNumberEquals } =
+          prepareEventsEmittedCounter(
+            UserStorageMockttpControllerEvents.PUT_SINGLE,
+          );
+
+        // Add a new account
+        await accountListPage.addAccount({
+          accountType: ACCOUNT_TYPE.Ethereum,
+        });
+
+        // Wait for sync operation to complete
+        await waitUntilSyncedAccountsNumberEquals(2);
+        await waitUntilEventsEmittedNumberEquals(1);
+
+        // Reopen account menu to verify both accounts are visible
+        await header.openAccountMenu();
+        await accountListPage.check_pageIsLoaded();
+        await accountListPage.check_accountDisplayedInAccountList(
+          DEFAULT_ACCOUNT_NAME,
+        );
+        await accountListPage.check_accountDisplayedInAccountList(
+          ADDED_ACCOUNT_NAME,
+        );
+
+        await accountListPage.closeAccountModal();
+      },
+    );
+
+    // Phase 2: Login to fresh instance, verify account persists, rename and add more accounts
+    await withFixtures(
+      {
+        fixtures: new FixtureBuilder()
+          .withBackupAndSyncSettings()
+          .build(),
+        title: this.test?.fullTitle(),
+        testSpecificMock: sharedMockSetup,
+      },
+      async ({ driver }) => {
+        await unlockWallet(driver);
+
+        const header = new HeaderNavbar(driver);
+        await header.check_pageIsLoaded();
+        await header.openAccountMenu();
+
+        const accountListPage = new AccountListPage(driver);
+        await accountListPage.check_pageIsLoaded();
+
+        // Verify both accounts from previous phase are still visible
+        await accountListPage.check_accountDisplayedInAccountList(
+          DEFAULT_ACCOUNT_NAME,
+        );
+        await accountListPage.check_accountDisplayedInAccountList(
+          ADDED_ACCOUNT_NAME,
+        );
+
+                // Set up event counter to track sync operations
+        const { prepareEventsEmittedCounter, waitUntilSyncedAccountsNumberEquals } = arrangeTestUtils(
+          driver,
+          userStorageMockttpController,
+        );
+        const { waitUntilEventsEmittedNumberEquals } =
+          prepareEventsEmittedCounter(
+            UserStorageMockttpControllerEvents.PUT_SINGLE,
+          );
+
+        // Rename the second account
+        await accountListPage.openAccountDetailsModal(ADDED_ACCOUNT_NAME);
+        const accountDetailsModal = new AccountDetailsModal(driver);
+        await accountDetailsModal.check_pageIsLoaded();
+        await accountDetailsModal.changeAccountLabel(NEW_ACCOUNT_NAME);
+
+        // Reopen account menu to add a third account
+        await header.openAccountMenu();
+        await accountListPage.check_pageIsLoaded();
+
+        // Add a third account
+        await accountListPage.addAccount({
+          accountType: ACCOUNT_TYPE.Ethereum,
+        });
+
+        // Wait for both sync operations to complete (rename + add)
+        await waitUntilSyncedAccountsNumberEquals(3);
+        await waitUntilEventsEmittedNumberEquals(2);
+
+        // Reopen account menu to verify all accounts are visible with correct names
+        await header.openAccountMenu();
+        await accountListPage.check_pageIsLoaded();
+        await accountListPage.check_accountDisplayedInAccountList(
+          DEFAULT_ACCOUNT_NAME,
+        );
+        await accountListPage.check_accountDisplayedInAccountList(
+          NEW_ACCOUNT_NAME,
+        );
+        await accountListPage.check_accountDisplayedInAccountList(
+          LAST_ACCOUNT_NAME,
+        );
+
+        await accountListPage.closeAccountModal();
+      },
+    );
+
+    // Phase 3: Complete onboarding flow from scratch to verify all changes persist
+    await withFixtures(
+      {
+        fixtures: new FixtureBuilder({ onboarding: true }).build(),
+        title: this.test?.fullTitle(),
+        testSpecificMock: sharedMockSetup,
+      },
+      async ({ driver }) => {
+        // Go through onboarding again to ensure accounts and names are synced
+        await completeImportSRPOnboardingFlow({
+          driver,
+          seedPhrase: E2E_SRP,
+          password: WALLET_PASSWORD,
+        });
+
+        // Wait for sync to complete during onboarding
+        await driver.delay(2000);
+
+        const header = new HeaderNavbar(driver);
+        await header.check_pageIsLoaded();
+        await header.openAccountMenu();
+
+        const accountListPage = new AccountListPage(driver);
+        await accountListPage.check_pageIsLoaded();
+
+        // Verify all accounts and renames are properly synced
+        await accountListPage.check_accountDisplayedInAccountList(
+          DEFAULT_ACCOUNT_NAME,
+        );
+        await accountListPage.check_accountDisplayedInAccountList(
+          NEW_ACCOUNT_NAME,
+        );
+        await accountListPage.check_accountDisplayedInAccountList(
+          LAST_ACCOUNT_NAME,
+        );
+
+        // Note: We don't check the exact count since onboarding with E2E_SRP
+        // may discover additional accounts with balances/activity on-chain
+        // The important thing is that our synced accounts are present
+      },
+    );
+  });
+});
