--- test/e2e/tests/identity/account-syncing/sync-after-adding-account.spec.ts
@@ -1,261 +1,0 @@
-import { Mockttp } from 'mockttp';
-import { USER_STORAGE_FEATURE_NAMES } from '@metamask/profile-sync-controller/sdk';
-import { withFixtures, unlockWallet } from '../../../helpers';
-import FixtureBuilder from '../../../fixture-builder';
-import { mockIdentityServices } from '../mocks';
-import { ACCOUNT_TYPE } from '../../../constants';
-import {
-  UserStorageMockttpController,
-  UserStorageMockttpControllerEvents,
-} from '../../../helpers/identity/user-storage/userStorageMockttpController';
-import HeaderNavbar from '../../../page-objects/pages/header-navbar';
-import AccountListPage from '../../../page-objects/pages/account-list-page';
-import HomePage from '../../../page-objects/pages/home/homepage';
-import {
-  accountsToMockForAccountsSync,
-  getAccountsSyncMockResponse,
-} from './mock-data';
-import { arrangeTestUtils } from './helpers';
-
-describe('Account syncing - Add Account', function () {
-  this.timeout(160000); // This test is very long, so we need an unusually high timeout
-
-  const arrange = async () => {
-    const unencryptedAccounts = accountsToMockForAccountsSync;
-    const mockedAccountSyncResponse = await getAccountsSyncMockResponse();
-    const customNameAccount3 = '3rd Account';
-    const defaultNameAccount3 = 'Account 3';
-
-    const userStorageMockttpController = new UserStorageMockttpController();
-
-    return {
-      unencryptedAccounts,
-      mockedAccountSyncResponse,
-      userStorageMockttpController,
-      customNameAccount3,
-      defaultNameAccount3,
-    };
-  };
-
-  describe('from inside MetaMask', function () {
-    it('syncs newly added accounts - custom name', async function () {
-      const {
-        unencryptedAccounts,
-        mockedAccountSyncResponse,
-        userStorageMockttpController,
-        customNameAccount3,
-      } = await arrange();
-
-      await withFixtures(
-        {
-          fixtures: new FixtureBuilder()
-            .withBackupAndSyncSettings()
-            .build(),
-          title: this.test?.fullTitle(),
-          testSpecificMock: (server: Mockttp) => {
-            userStorageMockttpController.setupPath(
-              USER_STORAGE_FEATURE_NAMES.accounts,
-              server,
-              {
-                getResponse: mockedAccountSyncResponse,
-              },
-            );
-
-            return mockIdentityServices(server, userStorageMockttpController);
-          },
-        },
-        async ({ driver }) => {
-          await unlockWallet(driver);
-
-          const header = new HeaderNavbar(driver);
-          await header.check_pageIsLoaded();
-          await header.openAccountMenu();
-
-          const accountListPage = new AccountListPage(driver);
-          await accountListPage.check_pageIsLoaded();
-          await accountListPage.check_numberOfAvailableAccounts(
-            mockedAccountSyncResponse.length,
-            ACCOUNT_TYPE.Ethereum,
-          );
-          await accountListPage.check_accountDisplayedInAccountList(
-            unencryptedAccounts[0].n,
-          );
-          await accountListPage.check_accountDisplayedInAccountList(
-            unencryptedAccounts[1].n,
-          );
-
-          const {
-            prepareEventsEmittedCounter,
-            waitUntilSyncedAccountsNumberEquals,
-          } = arrangeTestUtils(driver, userStorageMockttpController);
-
-          const { waitUntilEventsEmittedNumberEquals } =
-            prepareEventsEmittedCounter(
-              UserStorageMockttpControllerEvents.PUT_SINGLE,
-            );
-
-          await accountListPage.addAccount({
-            accountType: ACCOUNT_TYPE.Ethereum,
-            accountName: customNameAccount3,
-          });
-
-          // Wait for the account AND account name to be synced
-          await waitUntilSyncedAccountsNumberEquals(3);
-          await waitUntilEventsEmittedNumberEquals(1);
-        },
-      );
-
-      await withFixtures(
-        {
-          fixtures: new FixtureBuilder()
-            .withBackupAndSyncSettings()
-            .build(),
-          title: this.test?.fullTitle(),
-          testSpecificMock: (server: Mockttp) => {
-            userStorageMockttpController.setupPath(
-              USER_STORAGE_FEATURE_NAMES.accounts,
-              server,
-            );
-            return mockIdentityServices(server, userStorageMockttpController);
-          },
-        },
-        async ({ driver }) => {
-          await unlockWallet(driver);
-
-          const header = new HeaderNavbar(driver);
-          await header.check_pageIsLoaded();
-          await header.openAccountMenu();
-
-          const accountListPage = new AccountListPage(driver);
-          await accountListPage.check_pageIsLoaded();
-
-          const accountSyncResponse = userStorageMockttpController.paths.get(
-            USER_STORAGE_FEATURE_NAMES.accounts,
-          )?.response;
-
-          await accountListPage.check_numberOfAvailableAccounts(
-            accountSyncResponse?.length as number,
-            ACCOUNT_TYPE.Ethereum,
-          );
-          await accountListPage.check_accountDisplayedInAccountList(
-            unencryptedAccounts[0].n,
-          );
-          await accountListPage.check_accountDisplayedInAccountList(
-            unencryptedAccounts[1].n,
-          );
-          await accountListPage.check_accountDisplayedInAccountList(
-            customNameAccount3,
-          );
-        },
-      );
-    });
-
-    it('syncs newly added accounts - default name', async function () {
-      const {
-        unencryptedAccounts,
-        mockedAccountSyncResponse,
-        userStorageMockttpController,
-        defaultNameAccount3,
-      } = await arrange();
-
-      await withFixtures(
-        {
-          fixtures: new FixtureBuilder()
-            .withBackupAndSyncSettings()
-            .build(),
-          title: this.test?.fullTitle(),
-          testSpecificMock: (server: Mockttp) => {
-            userStorageMockttpController.setupPath(
-              USER_STORAGE_FEATURE_NAMES.accounts,
-              server,
-              {
-                getResponse: mockedAccountSyncResponse,
-              },
-            );
-
-            return mockIdentityServices(server, userStorageMockttpController);
-          },
-        },
-        async ({ driver }) => {
-          await unlockWallet(driver);
-
-          const header = new HeaderNavbar(driver);
-          await header.check_pageIsLoaded();
-          await header.openAccountMenu();
-
-          const accountListPage = new AccountListPage(driver);
-          await accountListPage.check_pageIsLoaded();
-          await accountListPage.check_numberOfAvailableAccounts(
-            mockedAccountSyncResponse.length,
-            ACCOUNT_TYPE.Ethereum,
-          );
-          await accountListPage.check_accountDisplayedInAccountList(
-            unencryptedAccounts[0].n,
-          );
-          await accountListPage.check_accountDisplayedInAccountList(
-            unencryptedAccounts[1].n,
-          );
-
-          const { waitUntilSyncedAccountsNumberEquals } = arrangeTestUtils(
-            driver,
-            userStorageMockttpController,
-          );
-
-          await accountListPage.addAccount({
-            accountType: ACCOUNT_TYPE.Ethereum,
-          });
-
-          // Wait for the account to be synced
-          await waitUntilSyncedAccountsNumberEquals(3);
-        },
-      );
-
-      await withFixtures(
-        {
-          fixtures: new FixtureBuilder()
-            .withBackupAndSyncSettings()
-            .build(),
-          title: this.test?.fullTitle(),
-          testSpecificMock: (server: Mockttp) => {
-            userStorageMockttpController.setupPath(
-              USER_STORAGE_FEATURE_NAMES.accounts,
-              server,
-            );
-            return mockIdentityServices(server, userStorageMockttpController);
-          },
-        },
-        async ({ driver }) => {
-          await unlockWallet(driver);
-          const homePage = new HomePage(driver);
-          await homePage.check_pageIsLoaded();
-          await homePage.check_expectedBalanceIsDisplayed('0');
-
-          const header = new HeaderNavbar(driver);
-          await header.check_pageIsLoaded();
-          await header.openAccountMenu();
-
-          const accountListPage = new AccountListPage(driver);
-          await accountListPage.check_pageIsLoaded();
-
-          const accountSyncResponse = userStorageMockttpController.paths.get(
-            USER_STORAGE_FEATURE_NAMES.accounts,
-          )?.response;
-
-          await accountListPage.check_numberOfAvailableAccounts(
-            accountSyncResponse?.length as number,
-            ACCOUNT_TYPE.Ethereum,
-          );
-          await accountListPage.check_accountDisplayedInAccountList(
-            unencryptedAccounts[0].n,
-          );
-          await accountListPage.check_accountDisplayedInAccountList(
-            unencryptedAccounts[1].n,
-          );
-          await accountListPage.check_accountDisplayedInAccountList(
-            defaultNameAccount3,
-          );
-        },
-      );
-    });
-  });
-});
